Ensure only one MintingUpdate event
Make sure that always one and only one MintingUpdate event is emitted when changing the position, including on clone, but not on the initial creation of the parent position. Also, the PositionEvent was adjusted to include the original, but not the price any more, as the price should be taken from the MintingUpdate event.